funcion eliminar en todo voluntarios

+++ Sistema/Voluntarios/Gestion_tbl_Voluntarios_P.php
@@ -111,7 +111,7 @@ if ($datos=$sql->fetch_object()) {
 }
 $sql=$conexion->query("SELECT * FROM tbl_permisos where Permiso_Eliminacion=1 and ID_Rol=$ID_Rol and ID_Objeto=15");
 if ($datos=$sql->fetch_object()) { 
-        $output['data'] .= '<td><a class="boton-eliminar" href="Delete_voluntarios_proyectos.php?ID_Vinculacion_Proy=' . $row['ID_Vinculacion_Proy'] . '"><i class="zmdi zmdi-delete"></i></a></td>';
+        $output['data'] .= '<td><a onclick="return confirmar()" class="boton-eliminar" href="Delete_voluntarios_proyectos.php?ID_Vinculacion_Proy=' . $row['ID_Vinculacion_Proy'] . '"><i class="zmdi zmdi-delete"></i></a></td>';
 }
         $output['data'] .= '</tr>';
     }
